+++ packages/browser-wallet/src/popup/popupX/pages/About/About.tsx
@@ -2,14 +2,13 @@ import { useTranslation } from 'react-i18next';
 import Page from '@popup/popupX/shared/Page';
 import Button from '@popup/popupX/shared/Button';
 import ArrowSquare from '@assets/svgX/arrow-square-out.svg';
-import React, { useEffect } from 'react';
+import React from 'react';
 import Card from '@popup/popupX/shared/Card';
 import Text from '@popup/popupX/shared/Text';
 import { getVersionName } from '@shared/utils/environment-helpers';
 import { useAtomValue } from 'jotai';
 import { acceptedTermsAtom } from '@popup/store/settings';
 import urls from '@shared/constants/url';
-import appTracker from '@shared/analytics';
 
 // Start saving the URL for the T&C we get from the wallet proxy,
 // and use that in the About page link.
@@ -27,10 +26,6 @@ export default function About() {
     const { t } = useTranslation('x', { keyPrefix: 'aboutPage' });
     const termsUrl = useGetTermsUrl();
 
-    useEffect(() => {
-        appTracker.aboutScreen();
-    }, []);
-
     return (
         <Page className="about-x">
             <Page.Top heading={t('about')} />
@@ -39,13 +34,7 @@ export default function About() {
                     <Card.Row>
                         <Text.MainMedium>{t('documentation')}</Text.MainMedium>
                         <Text.ExternalLink path={urls.documentationWebsite}>
-                            <Button.Base
-                                as="span"
-                                className="button__icon transparent"
-                                onClick={() => {
-                                    appTracker.aboutScreenLinkClicked(urls.documentationWebsite);
-                                }}
-                            >
+                            <Button.Base as="span" className="button__icon transparent">
                                 <ArrowSquare />
                             </Button.Base>
                         </Text.ExternalLink>
@@ -53,13 +42,7 @@ export default function About() {
                     <Card.Row>
                         <Text.MainMedium>{t('forum')}</Text.MainMedium>
                         <Text.ExternalLink path={urls.supportWebsite}>
-                            <Button.Base
-                                as="span"
-                                className="button__icon transparent"
-                                onClick={() => {
-                                    appTracker.aboutScreenLinkClicked(urls.supportWebsite);
-                                }}
-                            >
+                            <Button.Base as="span" className="button__icon transparent">
                                 <ArrowSquare />
                             </Button.Base>
                         </Text.ExternalLink>
@@ -67,13 +50,7 @@ export default function About() {
                     <Card.Row>
                         <Text.MainMedium>{t('website')}</Text.MainMedium>
                         <Text.ExternalLink path={urls.website}>
-                            <Button.Base
-                                as="span"
-                                className="button__icon transparent"
-                                onClick={() => {
-                                    appTracker.aboutScreenLinkClicked(urls.website);
-                                }}
-                            >
+                            <Button.Base as="span" className="button__icon transparent">
                                 <ArrowSquare />
                             </Button.Base>
                         </Text.ExternalLink>
@@ -81,13 +58,7 @@ export default function About() {
                     <Card.Row>
                         <Text.MainMedium>{t('termsAndConditions')}</Text.MainMedium>
                         <Text.ExternalLink path={termsUrl}>
-                            <Button.Base
-                                as="span"
-                                className="button__icon transparent"
-                                onClick={() => {
-                                    appTracker.aboutScreenLinkClicked(termsUrl);
-                                }}
-                            >
+                            <Button.Base as="span" className="button__icon transparent">
                                 <ArrowSquare />
                             </Button.Base>
                         </Text.ExternalLink>
@@ -95,13 +66,7 @@ export default function About() {
                     <Card.Row>
                         <Text.MainMedium>{t('licence')}</Text.MainMedium>
                         <Text.ExternalLink path={urls.licenseAttributions}>
-                            <Button.Base
-                                as="span"
-                                className="button__icon transparent"
-                                onClick={() => {
-                                    appTracker.aboutScreenLinkClicked(urls.licenseAttributions);
-                                }}
-                            >
+                            <Button.Base as="span" className="button__icon transparent">
                                 <ArrowSquare />
                             </Button.Base>
                         </Text.ExternalLink>