+++ packages/browser-wallet/src/popup/popupX/pages/IdIssuance/Submitted.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAtomValue } from 'jotai';
 
 import Button from '@popup/popupX/shared/Button';
-import { ConfirmedIdCard, PendingIdCard, RejectedIdCard } from '@popup/popupX/shared/IdCard';
+import { ConfirmedIdCard, RejectedIdCard, PendingIdCard } from '@popup/popupX/shared/IdCard';
 import Page from '@popup/popupX/shared/Page';
 import Text from '@popup/popupX/shared/Text';
 import { identitiesAtomWithLoading } from '@popup/store/identity';
@@ -11,22 +11,11 @@ import { CreationStatus } from '@shared/storage/types';
 import { absoluteRoutes } from '@popup/popupX/constants/routes';
 import { isSpawnedWindow } from '@popup/shared/window-helpers';
 import { useNavigate } from 'react-router-dom';
-import appTracker from '@shared/analytics';
 
 export default function IdIssuanceSubmitted() {
     const { t } = useTranslation('x', { keyPrefix: 'idIssuance.submitted' });
     const { loading, value: identities } = useAtomValue(identitiesAtomWithLoading);
     const nav = useNavigate();
-
-    useEffect(() => {
-        appTracker.identityVerificationResultScreen();
-    }, []);
-
-    if (loading) {
-        return null;
-    }
-
-    const identity = identities.slice(-1)[0];
 
     const handleDone = () => {
         if (isSpawnedWindow) {
@@ -34,11 +23,13 @@ export default function IdIssuanceSubmitted() {
         } else {
             nav(absoluteRoutes.home.path, { replace: true });
         }
-
-        if (identity.status === CreationStatus.Confirmed) {
-            appTracker.identityVerificationResultApprovedDialog();
-        }
     };
+
+    if (loading) {
+        return null;
+    }
+
+    const identity = identities.slice(-1)[0];
 
     return (
         <Page>