+++ packages/browser-wallet/src/popup/popupX/pages/Onboarding/Welcome.tsx
@@ -1,11 +1,6 @@
-import React, { ReactNode, useEffect, useState } from 'react';
+import React, { ReactNode, useEffect } from 'react';
 import { useAtom } from 'jotai';
-import {
-    acceptedActivityTrackingAtom,
-    acceptedTermsAtom,
-    networkConfigurationAtom,
-    uiStyleAtom,
-} from '@popup/store/settings';
+import { acceptedTermsAtom, networkConfigurationAtom, uiStyleAtom } from '@popup/store/settings';
 import { mainnet } from '@shared/constants/networkConfiguration';
 import Shield from '@assets/svgX/crypto-currency-square-shield.svg';
 import Assets from '@assets/svgX/crypto-currency-assets.svg';
@@ -21,8 +16,6 @@ import Page from '@popup/popupX/shared/Page';
 import { useAsyncMemo } from 'wallet-common-helpers';
 import { getTermsAndConditionsConfig } from '@shared/utils/network-helpers';
 import { UiStyle } from '@shared/storage/types';
-import { Checkbox } from '@popup/popupX/shared/Form/Checkbox';
-import appTracker from '@shared/analytics';
 
 const bg = document.getElementsByClassName('bg').item(0);
 
@@ -42,29 +35,17 @@ export default function Welcome() {
     const { t } = useTranslation('x', { keyPrefix: 'onboarding.welcome' });
     const [, setCurrentNetworkConfiguration] = useAtom(networkConfigurationAtom);
     const [{ value: acceptedTerms }, setAcceptedTerms] = useAtom(acceptedTermsAtom);
-    const [, setActivityTracking] = useAtom(acceptedActivityTrackingAtom);
     const [, setUiStyle] = useAtom(uiStyleAtom);
-    const [checkboxTerms, setCheckboxTerms] = useState(false);
-    const [checkboxTracking, setCheckboxTracking] = useState(false);
     const config = useAsyncMemo(getTermsAndConditionsConfig, undefined, []);
     const nav = useNavigate();
-    const navToPassword = async () => {
+    const navToPassword = () => {
         const version = config?.version || acceptedTerms?.version;
         // If we didn't find a version, put in an empty version
-        setAcceptedTerms({ accepted: checkboxTerms, version: version || '', url: config?.url });
+        setAcceptedTerms({ accepted: true, version: version || '', url: config?.url });
 
         setUiStyle(UiStyle.WalletX);
 
-        try {
-            await setActivityTracking(appTracker.createAnalyticsTrackingObject(checkboxTracking));
-
-            appTracker.welcomeScreen();
-            appTracker.welcomeActivityTrackingCheckBoxChecked();
-            appTracker.welcomeTermAndConditionsCheckBoxChecked();
-            appTracker.welcomeGetStartedClicked();
-        } finally {
-            nav(absoluteRoutes.onboarding.setupPassword.path);
-        }
+        return nav(absoluteRoutes.onboarding.setupPassword.path);
     };
 
     useEffect(() => {
@@ -84,25 +65,19 @@ export default function Welcome() {
                 <Description icon={<Possibilities />} title={t('unlimited')} description={t('transactionInvest')} />
             </Page.Main>
             <Page.Footer>
-                <div className="welcome__check_group">
-                    <Text.Capture>
-                        <Checkbox checked={checkboxTerms} onChange={(e) => setCheckboxTerms(e.target.checked)} />
-                        {t('proceeding')}
-                        <ExternalLink path={urls.termsAndConditions}>{t('termsAndConditions')}</ExternalLink>
-                    </Text.Capture>
-                    <Text.Capture>
-                        <Checkbox checked={checkboxTracking} onChange={(e) => setCheckboxTracking(e.target.checked)} />
-                        {t('activityTracking')}
-                    </Text.Capture>
-                </div>
                 <Button.Main
                     label={t('start')}
                     onClick={() => {
                         bg?.classList.remove('welcome-page');
                         navToPassword();
                     }}
-                    disabled={!checkboxTerms}
                 />
+                <div className="welcome__footer">
+                    <Text.Capture>
+                        {t('proceeding')}
+                        <ExternalLink path={urls.termsAndConditions}>{t('termsAndConditions')}</ExternalLink>
+                    </Text.Capture>
+                </div>
             </Page.Footer>
         </Page>
     );