+++ packages/browser-wallet/src/popup/popupX/shared/PasswordSession/PasswordSession.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import { useAtom, useAtomValue, useSetAtom } from 'jotai';
+import { useAtomValue, useSetAtom } from 'jotai';
 import { useTranslation } from 'react-i18next';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { acceptedActivityTrackingAtom, encryptedSeedPhraseAtom, sessionPasscodeAtom } from '@popup/store/settings';
+import { encryptedSeedPhraseAtom, sessionPasscodeAtom } from '@popup/store/settings';
 import { decrypt } from '@shared/utils/crypto';
 import Page from '@popup/popupX/shared/Page';
 import Text from '@popup/popupX/shared/Text';
 import Form from '@popup/popupX/shared/Form/Form';
 import FormPassword from '@popup/popupX/shared/Form/Password';
 import Button from '@popup/popupX/shared/Button';
-import appTracker from '@shared/analytics';
 
 type FormValues = {
     passcode: string;
@@ -18,7 +17,6 @@ type FormValues = {
 export default function PasswordSession() {
     const setPasscodeInSession = useSetAtom(sessionPasscodeAtom);
     const encryptedSeedPhrase = useAtomValue(encryptedSeedPhraseAtom);
-    const [activityTracking, setActivityTracking] = useAtom(acceptedActivityTrackingAtom);
     const { t } = useTranslation('x', { keyPrefix: 'sharedX.passwordSession' });
     const { t: tPasscode } = useTranslation('x', { keyPrefix: 'sharedX.form.password' });
     const form = useForm<FormValues>();
@@ -28,9 +26,6 @@ export default function PasswordSession() {
     }
 
     const handleSubmit: SubmitHandler<FormValues> = async (vs) => {
-        if (activityTracking.value) {
-            setActivityTracking({ ...activityTracking.value, sessionId: appTracker.getDateTimestamp() });
-        }
         if (encryptedSeedPhrase.value) {
             try {
                 await decrypt(encryptedSeedPhrase.value, vs.passcode);